Split up Condition class

c
