+++ psalm-baseline.xml
@@ -113,9 +113,6 @@
     <RedundantCondition>
       <code>$state === 1</code>
     </RedundantCondition>
-    <UnusedParam>
-      <code>$options</code>
-    </UnusedParam>
   </file>
   <file src="src/Components/Expression.php">
     <MixedArrayOffset>
@@ -179,9 +176,6 @@
     <RedundantCondition>
       <code>$state === 1</code>
     </RedundantCondition>
-    <UnusedParam>
-      <code>$options</code>
-    </UnusedParam>
   </file>
   <file src="src/Components/GroupKeyword.php">
     <PossiblyNullPropertyAssignmentValue>
@@ -287,9 +281,6 @@
     <RedundantCondition>
       <code>$state === 5</code>
     </RedundantCondition>
-    <UnusedParam>
-      <code>$options</code>
-    </UnusedParam>
   </file>
   <file src="src/Components/Limit.php">
     <MixedAssignment>
@@ -310,9 +301,6 @@
     <PossiblyNullPropertyAssignmentValue>
       <code><![CDATA[Expression::parse($parser, $list, ['parseField' => 'table'])]]></code>
     </PossiblyNullPropertyAssignmentValue>
-    <UnusedParam>
-      <code>$options</code>
-    </UnusedParam>
   </file>
   <file src="src/Components/OptionsArray.php">
     <MixedArgument>
@@ -475,9 +463,6 @@
     <RedundantConditionGivenDocblockType>
       <code><![CDATA[! empty($this->options)]]></code>
     </RedundantConditionGivenDocblockType>
-    <UnusedParam>
-      <code>$options</code>
-    </UnusedParam>
     <UnusedVariable>
       <code>$idx</code>
     </UnusedVariable>
@@ -501,9 +486,6 @@
     <RedundantCondition>
       <code>$state === 2</code>
     </RedundantCondition>
-    <UnusedParam>
-      <code>$options</code>
-    </UnusedParam>
   </file>
   <file src="src/Components/RenameOperation.php">
     <PossiblyNullPropertyAssignmentValue>
