+++ src/Components/CaseExpression.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace PhpMyAdmin\SqlParser\Components;
 
 use PhpMyAdmin\SqlParser\Component;
+use PhpMyAdmin\SqlParser\Components\Lists\Conditions;
 use PhpMyAdmin\SqlParser\Context;
 use PhpMyAdmin\SqlParser\Parser;
 use PhpMyAdmin\SqlParser\Token;
@@ -104,7 +105,7 @@ final class CaseExpression implements Component
                     switch ($token->keyword) {
                         case 'WHEN':
                             ++$list->idx; // Skip 'WHEN'
-                            $newCondition = Condition::parse($parser, $list);
+                            $newCondition = Conditions::parse($parser, $list);
                             $type = 1;
                             $state = 1;
                             $ret->conditions[] = $newCondition;
@@ -264,7 +265,7 @@ final class CaseExpression implements Component
             $valuesCount = count($this->conditions);
             $resultsCount = count($this->results);
             for ($i = 0; $i < $valuesCount && $i < $resultsCount; ++$i) {
-                $ret .= 'WHEN ' . Condition::buildAll($this->conditions[$i]) . ' ';
+                $ret .= 'WHEN ' . Conditions::buildAll($this->conditions[$i]) . ' ';
                 $ret .= 'THEN ' . $this->results[$i] . ' ';
             }
         }
